requirements.txt
flask
rembg
pillow
opencv-python-headless
numpy
gunicorn
